+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 Unreleased
 ----------
 
+- Add RawKey for key material. `#138 <https://github.com/dajiaji/python-cwt/pull/138>`__
 - Make MAC key can be used for KDF. `#137 <https://github.com/dajiaji/python-cwt/pull/137>`__
 - Remove COSEKeyInterface from RecipientInterface. `#137 <https://github.com/dajiaji/python-cwt/pull/137>`__
 - Implement AESKeyWrap which has COSEKeyInterface. `#137 <https://github.com/dajiaji/python-cwt/pull/137>`__